Update sign in and sign up pages to use supabase
