+++ frontend/Lingual_App/app/(auth)/sign-up.jsx
@@ -2,6 +2,7 @@ import { View, Text, ScrollView, Image } from 'react-native'
 import React, { useState } from 'react'
 import { SafeAreaView } from 'react-native-safe-area-context'
 import { Link } from 'expo-router'
+import { supabase } from '../../utils/supabase'
 
 import { images } from '../../constants'
 
@@ -11,17 +12,33 @@ import IconButton from '../../components/iconButton'
 
 const SignUp = () => {
 
-    const [form, setform] = useState({
-        username: '',
-        email: '',
-        password: ''
-    })
-
     const [isSubmitting, setisSubmitting] = useState(false)
 
     const submit = () => {
 
     }
+
+    // Supabase Auth
+    const [email, setEmail] = useState('')
+    const [password, setPassword] = useState('')
+    const [loading, setLoading] = useState(false)
+
+    async function signUpWithEmail() {
+        setLoading(true)
+        const {
+            data: { session },
+            error,
+        } = await supabase.auth.signUp({
+            email: email,
+            password: password,
+        })
+
+        if (error) Alert.alert(error.message)
+        if (!session) Alert.alert('Please check your inbox for email verification!')
+        setLoading(false)
+    }
+
+    // End Supabase Auth
 
     return (
         <SafeAreaView className="bg-primary h-full">
@@ -38,25 +55,27 @@ const SignUp = () => {
                     {/* Email form field */}
                     <FormField
                         title="Email"
-                        value={form.email}
-                        handleChangeText={(e) => setform({ ...form, email: e })}
+                        value={email}
+                        handleChangeText={(text) => setEmail(text)}
                         otherStyles="mt-7"
                         keyboardType="email-address"
+                        placeholder='email@address.com'
                     />
 
                     {/* Password form field */}
                     <FormField
                         title="Password"
-                        value={form.password}
-                        handleChangeText={(e) => setform({ ...form, password: e })}
+                        value={password}
+                        handleChangeText={(text) => setPassword(text)}
                         otherStyles="mt-7"
+                        placeholder='Password'
                     />
 
                     <CustomButton
-                        title="Sign-Up"
-                        handlePress={submit}
+                        title="Sign-In"
+                        handlePress={() => signUpWithEmail()}
                         containerStyles="mt-7"
-                        isLoading={isSubmitting}
+                        isLoading={loading}
                     />
 
                     <View className="justify-center pt-5 flex-row gap-2">
@@ -87,14 +106,14 @@ const SignUp = () => {
                     <View className="pt-5 flex-row gap-2 items-center justify-evenly">
                         <IconButton
                             image={images.googleLogo}
-                            handlePress={submit}
+                            // handlePress={submit}
                             isLoading={isSubmitting}
                             containerStyles="w-16 h-16 rounded-full justify-center items-center border-2 border-black-100"
                         />
 
                         <IconButton
                             image={images.appleLogo}
-                            handlePress={submit}
+                            // handlePress={submit}
                             isLoading={isSubmitting}
                             containerStyles="w-16 h-16 rounded-full justify-center items-center border-2 border-black-100"
                         />
